requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
flask-cors==4.0.0
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
email-validator==2.1.0
python-jose==3.3.0
bcrypt==4.0.1
gunicorn==21.2.0
requests==2.31.0
tqdm==4.66.1
